src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
fastapi==0.104.0
h11==0.14.0
httptools==0.6.1
idna==3.4
motor==3.3.1
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
